modules/location-ingestion/requirements.txt
Flask==1.1.1
kafka-python==2.0.2
